refactor(renderer): extract the runtime offline visual state rule from the member helpers
`memberHelpers.ts` sits at its frozen cap of 1756 lines with no headroom, and
the runtime-offline rule is the part of it that is about to change.

`getCurrentRuntimeOfflineVisualState` and the Codex-native offline predicate it
delegates to move into `memberRuntimeVisualState.ts`, together with
`MemberLaunchVisualState`, which is the vocabulary that rule produces. The two
predicates `shouldDisplayMemberCurrentTask` still calls -
`isCodexNativeProcessTeammate` and `hasLiveRuntimeProcessEvidence` - are
exported from the new module; the rest stays module-private, as it was.

`memberHelpers.ts` re-exports the type so nothing outside changes, and the
dependency runs one way: the helpers import the rule, never the reverse.

No behaviour change: the moved code is byte-identical apart from the added
export keywords, and the existing tests pass unchanged.
